requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11 